Move World DSL to its module and re-export from DSL.hs
