+++ src/swarm-scenario/Swarm/Game/State/Landscape.hs
@@ -53,8 +53,7 @@ import Swarm.Game.Terrain (TerrainType (..), terrainIndexByName)
 import Swarm.Game.Universe as U
 import Swarm.Game.World
 import Swarm.Game.World.Coords
-import Swarm.Game.World.Eval (runWorld)
-import Swarm.Game.World.Gen (Seed)
+import Swarm.Game.World.DSL (runWorld)
 import Swarm.Util.Erasable
 import Swarm.Util.Lens (makeLensesNoSigs)
