+++ src/swarm-scenario/Swarm/Game/World/Render.hs
@@ -47,7 +47,7 @@ import Swarm.Game.Scenario.Topography.WorldDescription
 import Swarm.Game.State.Landscape
 import Swarm.Game.Universe
 import Swarm.Game.World.Coords
-import Swarm.Game.World.Gen (Seed)
+import Swarm.Game.World.DSL.Gen (Seed)
 import Swarm.Pretty (prettyText)
 import Swarm.Util (failT)
 import Swarm.Util.Content
